Improve robustness of test suite
In addition to specifying the total number of expected lint errors
in each fixture, specify line numbers and rules for each. This
ensures that the build will still break if the config is modified
in a way that happens to both introduce and remove a lint error
from any of the fixtures.

This also fixes a test error introduced in commit
a9d57274e89d81c547f74f853d3565267328b61c